+++ scout/server/blueprints/clinvar/controllers.py
@@ -353,9 +353,9 @@ def json_api_submission(submission_id):
     """
     variant_data = store.clinvar_objs(submission_id, "variant_data")
     obs_data = store.clinvar_objs(submission_id, "case_data")
-    if not variant_data or not obs_data:
-        flash("Submission must contain both Variant and CaseData info", "warning")
-        return
+
+    if None in [variant_data, obs_data]:
+        return (400, "Submission must contain both Variant and CaseData info")
 
     # Retrieve eventual assertion criteria for the submission
     extra_params = store.clinvar_assertion_criteria(variant_data[0]) or {}